+++ pkg/dbutil/table.go
@@ -22,7 +22,6 @@ import (
 	"github.com/pingcap/errors"
 	"github.com/pingcap/parser/ast"
 	"github.com/pingcap/parser/model"
-	"github.com/pingcap/tidb-tools/pkg/utils"
 	"github.com/pingcap/tidb/ddl"
 	_ "github.com/pingcap/tidb/planner/core"        // to setup expression.EvalAstExpr. See: https://github.com/pingcap/tidb/blob/a94cff903cd1e7f3b050db782da84273ef5592f4/planner/core/optimizer.go#L202
 	_ "github.com/pingcap/tidb/types/parser_driver" // for parser driver
@@ -35,7 +34,7 @@ func GetTableInfo(ctx context.Context, db *sql.DB, schemaName string, tableName 
 		return nil, errors.Trace(err)
 	}
 
-	ansiQuotes, err := utils.HasAnsiQuotesMode(db)
+	ansiQuotes, err := HasAnsiQuotesMode(db)
 	if err != nil {
 		return nil, errors.Trace(err)
 	}
